refactor(commerce): streamline create-order-checkout component
Simplify the checkout form by removing unused imports, replacing manual form state with a structured IShippingInfo type, and integrating `ProductsDetailsAndDiscounts` and `OrderShipmentConfirm` sub-components. This improves maintainability and prepares for future payment/shipment confirmation flows.
