+++ src/modules/commerce/components/create-order-checkout/products-details-and-discounts/products-details-and-discounts.tsx
