--- src/modules/commerce/components/create-order-checkoutNEW/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./create-order-checkout";
